requirements.txt
streamlit==1.40.2
google-generativeai==0.8.3
speechrecognition==3.11.0
gtts==2.5.4
pygame==2.6.1
Pillow==11.0.0
python-dotenv==1.0.1